Trying to improve performance of the spectral dimension

+++ src/moves/Move.cpp
@@ -9,6 +9,6 @@ Move::~Move() {
 
 double Move::getMoveProbability(double lambda, double alpha) {
     // note that alpha should be positive
-    return exp(-lambda / 4.0 * (sqrt(4 * alpha - 1) * Ntts +
-            sqrt(alpha * (4 - alpha)) * Ntss));
+    return exp(-lambda / 4.0 * (sqrt(4.0 * alpha - 1.0) * (double)Ntts +
+            sqrt(alpha * (4.0 - alpha)) * (double)Ntss));
 }
